% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_from_text.R
\name{adjacency_from_text}
\alias{adjacency_from_text}
\title{Create adjacency matrices from graph6, sparse6, or digraph6 symbols}
\usage{
adjacency_from_text(object, ...)
}
\arguments{
\item{object}{character vector with a mixture of graph6, sparse6 or digraph6
symbols}
}
\value{
A list of adjacency matrices.
}
\description{
Create adjacency matrices from graph6, sparse6, or digraph6 symbols
}
\details{
If \code{object} contains sparse6 symbols, which are in fact encoded
edgelists, the function will return corresponding adjacency matrices
creating temporary igraph objects internally.
}
