% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_adjacency.R
\name{as_adjacency}
\alias{as_adjacency}
\alias{as_adjacency.default}
\alias{as_adjacency.graph6}
\title{Convert a graph6 symbol to adjacency matrix}
\usage{
as_adjacency(object)

\method{as_adjacency}{default}(object)

\method{as_adjacency}{graph6}(object)
}
\arguments{
\item{object}{character, a graph6 symbol}
}
\value{
A square symmetric adjacency matrix representing the graph.
}
\description{
This function takes a graph6 symbol and converts it to a symmetric square
adjacency matrix.
}
\note{
This function aim at implementing format and methods originally
created by Brendan McKay, \email{bdm@cs.anu.edu.au} and
\url{http://cs.anu.edu.au/~bdm}.
}
\references{
\url{http://cs.anu.edu.au/people/bdm/data/formats.txt}
}
