% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dgraph6.R
\name{as_dgraph6}
\alias{as_dgraph6}
\alias{as_dgraph6.default}
\alias{as_dgraph6.matrix}
\alias{as_dgraph6.list}
\alias{as_dgraph6.igraph}
\alias{as_dgraph6.network}
\title{Convert adjacency matrix to dgraph6 symbols}
\usage{
as_dgraph6(object)

\method{as_dgraph6}{default}(object)

\method{as_dgraph6}{matrix}(object)

\method{as_dgraph6}{list}(object)

\method{as_dgraph6}{igraph}(object)

\method{as_dgraph6}{network}(object)
}
\arguments{
\item{object}{a square binary adjacency matrix or a list of thereof

See \link{rgraph6} for dgraph6 format description.}
}
\value{
A character vector of dgraph6 symbols.
}
\description{
This function converts a graph(s) as adjacency matrices to a dgraph6 symbol(s).
}
