% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_digraph6.R
\name{as_digraph6}
\alias{as_digraph6}
\alias{as_digraph6.matrix}
\alias{as_digraph6.igraph}
\alias{as_digraph6.network}
\alias{as_digraph6.list}
\alias{as_digraph6.default}
\title{Encode network data as \code{digraph6} symbols}
\usage{
as_digraph6(object)

\method{as_digraph6}{matrix}(object)

\method{as_digraph6}{igraph}(object)

\method{as_digraph6}{network}(object)

\method{as_digraph6}{list}(object)

\method{as_digraph6}{default}(object)
}
\arguments{
\item{object}{a matrix, an igraph object or a network object or a list
thereof. See Methods section below.}
}
\value{
A character vector of digraph6 symbols.
}
\description{
Given (a list of) adjacency matrices or igraph objects or network objects
return a vector of digraph6 symbols encoding them.
}
\details{
See \link{rgraph6} for digraph6 format description.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: If \code{object} is a matrix it is interpreted as an
adjacency matrix of a directed graph.

\item \code{igraph}: Igraph \code{object} needs to be a directed.

\item \code{network}: Network \code{object} needs to be directed.

\item \code{list}: If \code{object} is a list the function is applied to each
element.

\item \code{default}: Throws an error about the unhandled class.
}}

\examples{
# From adjacency matrix ----------------------------------------------------
am <- matrix(c(
  0,1,0,
  0,0,1,
  1,0,0),
  byrow=TRUE, ncol=3, nrow=3)
as_digraph6(am)

# From igraph objects ------------------------------------------------------
if(requireNamespace("igraph", quietly=TRUE)) {
  g <- igraph::graph_from_adjacency_matrix(am)
  as_digraph6(g)
}

# From network objects -----------------------------------------------------
if(requireNamespace("network", quietly=TRUE)) {
  net <- network::network(am)
  as_digraph6(net)
}

}
