% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_elist.R
\name{as_elist}
\alias{as_elist}
\alias{as_elist.default}
\alias{as_elist.graph6}
\alias{as_elist.dgraph6}
\alias{as_elist.sparse6}
\title{Convert sparse6 symbols to edgelists}
\usage{
as_elist(object)

\method{as_elist}{default}(object)

\method{as_elist}{graph6}(object)

\method{as_elist}{dgraph6}(object)

\method{as_elist}{sparse6}(object)
}
\arguments{
\item{object}{character vector of sparse6 symbols}
}
\value{
A list of length \code{length(object)} of edgelists corresponding to input sparse6 symbols.
}
\description{
This function takes a vector of sparse6 symbols and converts them into
undirected edgelists.
}
\references{
\url{http://cs.anu.edu.au/people/bdm/data/formats.txt}
}
