% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_igraph.R
\name{as_igraph}
\alias{as_igraph}
\title{Coonvert graph6 symbols to igraph objects}
\usage{
as_igraph(object, ...)
}
\arguments{
\item{object}{vector of graph6 symbols}

\item{...}{other arguments passed to \code{\link[igraph:graph_from_adjacency_matrix]{igraph::graph_from_adjacency_matrix()}}}
}
\value{
A list of igraph objects.
}
\description{
Coonvert graph6 symbols to igraph objects
}
