% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_network.R
\name{as_network}
\alias{as_network}
\title{Convert graph6 symbols to network objects}
\usage{
as_network(object, ...)
}
\arguments{
\item{object}{vector of graph6 symbols}

\item{...}{other arguments passed to \link[network:network]{network::as.network}}
}
\value{
A list of network objects.
}
\description{
Convert graph6 symbols to network objects
}
