% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sparse6.R
\name{as_sparse6}
\alias{as_sparse6}
\alias{as_sparse6.default}
\alias{as_sparse6.matrix}
\alias{as_sparse6.list}
\alias{as_sparse6.igraph}
\alias{as_sparse6.network}
\title{Convert edgelist to sparse6 symbols}
\usage{
as_sparse6(object)

\method{as_sparse6}{default}(object)

\method{as_sparse6}{matrix}(object)

\method{as_sparse6}{list}(object)

\method{as_sparse6}{igraph}(object)

\method{as_sparse6}{network}(object)
}
\arguments{
\item{object}{an edgelist, igraph, network object or a list thereof

See \link{rgraph6} for sparse6 format description.}
}
\value{
A character vector with sparse6 symbols.
}
\description{
This function converts a graph(s) edgelist to a sparse6 symbol(s).
}
