% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_format.R
\name{choose_format}
\alias{choose_format}
\alias{choose_format.default}
\alias{choose_format.list}
\title{Choose format heuristically}
\usage{
choose_format(object, ...)

\method{choose_format}{default}(object, ...)

\method{choose_format}{list}(object, ...)
}
\arguments{
\item{object}{Igraph/network object or list thereof}

\item{...}{other arguments, currently ignored}
}
\value{
Character value out of 'graph6', 'sparse6' or 'digraph6'. If \code{object}
is a list, a vector of such values of the length equal to the length of
\code{object}.
}
\description{
Given a graph suggest efficient format out of 'graph6', 'sparse6' or 'digraph6'.
}
\details{
If \code{object} is directed, the suggested format is 'digraph6'. If
\code{object} is undirected the function suggests 'sparse6' if density is less
than 0.15 and 'graph6' otherwise. This rule is approximate and will be
replaced with exact one in the future.
}
