% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decbin.R
\name{d2b}
\alias{d2b}
\title{Converting between decimal and binary numbers}
\usage{
d2b(dec, as = c("character", "numeric"))
}
\arguments{
\item{dec}{single decimal number to be converted}

\item{as}{one of "character" or "numeric", type of value returned, see Value}
}
\value{
Function \code{d2b}, depending on \code{as}, returns a character vector or
list of numeric vectors. In both cases the length is equal to the length of 
\code{dec}.
}
\description{
Functions for converting decimal numbers to binary numbers and vice versa.
}
