% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decbin.R
\name{decbin}
\alias{decbin}
\alias{d2b}
\alias{b2d}
\alias{b2d.character}
\alias{b2d.list}
\title{Converting between decimal and binary numbers}
\usage{
d2b(dec, as = c("character", "numeric"))

b2d(bin)

\method{b2d}{character}(bin)

\method{b2d}{list}(bin)
}
\arguments{
\item{dec}{single decimal number to be converted}

\item{as}{one of "character" or "numeric", type of value returned, see Value}

\item{bin}{either a character vector of strings or a list of numeric vectors.
In both cases only 0s and 1s as elements are allowed.}
}
\value{
Function \code{d2b}, depending on \code{as}, returns a character vector or
list of numeric vectors. In both cases the length is equal to the length of 
\code{dec}.

Function \code{b2d} returns an integer vector of decimal numbers.
}
\description{
Functions for converting decimal numbers to binary numbers and vice versa.
}
