% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist_from_text.R
\name{edgelist_from_text}
\alias{edgelist_from_text}
\title{Create edgelist matrices from 'graph6', 'sparse6', or 'digraph6' symbols}
\usage{
edgelist_from_text(object, ...)
}
\arguments{
\item{object}{character vector with a mixture of 'graph6', 'sparse6' or
'digraph6' symbols}

\item{...}{other arguments, currently ignored}
}
\value{
A list of adjacency matrices.
}
\description{
Create edgelist matrices from 'graph6', 'sparse6', or 'digraph6' symbols
}
\details{
If \code{object} contains 'graph6' or 'digraph6' symbols, which are in
fact encoded adjacency matrices, the function will return corresponding
edgelist matrices creating temporary igraph objects internally.
}
