% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandToLength.R
\name{expand_to_length}
\alias{expand_to_length}
\title{Expand vector by adding values at the start or end}
\usage{
expand_to_length(x, l = ceiling(length(x)/6) * 6, what = 0,
  where = c("end", "start"))
}
\arguments{
\item{x}{sequence to be expanded}

\item{l}{target length}

\item{what}{what to replicate at the start or end}

\item{where}{either "start" or "end"}
}
\value{
A vector of length \code{l}.
}
\description{
Expand a sequence \code{x} to be of length \code{l} by adding \code{what} at
the start or end.
}
