% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_dgraph6.R
\name{from_dgraph6}
\alias{from_dgraph6}
\alias{adjacency_from_dgraph6}
\alias{igraph_from_dgraph6}
\alias{network_from_dgraph6}
\title{Functions parsing dgraph6 symbols}
\usage{
adjacency_from_dgraph6(d6)

igraph_from_dgraph6(d6, ...)

network_from_dgraph6(d6, ...)
}
\arguments{
\item{d6}{character vector of dgraph6 symbols}

\item{...}{other arguments, see Details.}
}
\value{
The returned object is:

\itemize{
\item for \code{\link[=adjacency_from_dgraph6]{adjacency_from_dgraph6()}}, a list of the same length as
its input of square symmetric adjacency matrices.
}

\itemize{
\item for \code{\link[=igraph_from_dgraph6]{igraph_from_dgraph6()}}, a list of igraph objects
}

\itemize{
\item for \code{\link[=network_from_dgraph6]{network_from_dgraph6()}}, a list of network objects
}
}
\description{
These functions take a vector of dgraph6 symbols and return a
list of other types of objects:

\itemize{
\item \code{\link[=adjacency_from_dgraph6]{adjacency_from_dgraph6()}} creates adjacency matrices
}

\itemize{
\item \code{\link[=igraph_from_dgraph6]{igraph_from_dgraph6()}} creates igraph objects. Requires
package \pkg{igraph} to be installed.
}

\itemize{
\item \code{\link[=network_from_dgraph6]{network_from_dgraph6()}} creates network objects. Requires
package \pkg{network} to be installed.
}
}
\details{
For \code{\link[=igraph_from_dgraph6]{igraph_from_dgraph6()}} additional arguments are passed to
\code{\link[igraph:graph_from_adjacency_matrix]{igraph::graph_from_adjacency_matrix()}}

For \code{\link[=network_from_dgraph6]{network_from_dgraph6()}} additional arguments are passed to
\code{\link[network:network]{network::as.network()}}
}
\seealso{
\code{\link[=as_dgraph6]{as_dgraph6()}} for saving objects as dgraph6 symbols.
}
