% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sparse6.R
\name{from_sparse6}
\alias{from_sparse6}
\alias{edgelist_from_sparse6}
\alias{igraph_from_sparse6}
\alias{network_from_sparse6}
\title{Functions parsing sparse6 symbols}
\usage{
edgelist_from_sparse6(s6)

igraph_from_sparse6(s6, ...)

network_from_sparse6(s6, ...)
}
\arguments{
\item{s6}{character vector of sparse6 symbols}

\item{...}{other arguments, see Details.}
}
\value{
The returned object is:

\itemize{
\item for \code{\link[=edgelist_from_sparse6]{edgelist_from_sparse6()}}, a list of the same length as
its input of two-column edgelist matrices.
}

\itemize{
\item for \code{\link[=igraph_from_sparse6]{igraph_from_sparse6()}}, a list of igraph objects
}

\itemize{
\item for \code{\link[=network_from_sparse6]{network_from_sparse6()}}, a list of network objects
}
}
\description{
These functions take a vector of sparse6 symbols and return a
list of other types of objects:

\itemize{
\item \code{\link[=edgelist_from_sparse6]{edgelist_from_sparse6()}} creates edgelist matrices
}

\itemize{
\item \code{\link[=igraph_from_sparse6]{igraph_from_sparse6()}} creates igraph objects. Requires
package \pkg{igraph} to be installed.
}

\itemize{
\item \code{\link[=network_from_sparse6]{network_from_sparse6()}} creates network objects. Requires
package \pkg{network} to be installed.
}
}
\details{
For \code{\link[=igraph_from_sparse6]{igraph_from_sparse6()}} additional arguments are passed to
\code{\link[igraph:graph_from_edgelist]{igraph::graph_from_edgelist()}}

For \code{\link[=network_from_sparse6]{network_from_sparse6()}} additional arguments are passed to
\code{\link[network:network]{network::as.network()}}
}
\seealso{
\code{\link[=as_sparse6]{as_sparse6()}} for saving objects as sparse6 symbols.
}
