% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_as_text.R
\name{graph_as_text}
\alias{graph_as_text}
\alias{graph_as_text.default}
\alias{graph_as_text.list}
\title{Encode graph as text}
\usage{
graph_as_text(object, ...)

\method{graph_as_text}{default}(object, ...)

\method{graph_as_text}{list}(object, ...)
}
\arguments{
\item{object}{igraph/network object or a list thereof}

\item{...}{other arguments, currently ignored}
}
\value{
A character vector of encoded graphs.
}
\description{
Encode a graph as 'graph6', 'sparse6' or 'digraph6' choosing the format
automatically.
}
\details{
If \code{object} is a list it may be a mixture of network and igraph
objects.
}
