% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_from_graph6.R
\name{igraph_from_graph6}
\alias{igraph_from_graph6}
\title{Convert graph6 symbols to igraph objects}
\usage{
igraph_from_graph6(object, ...)
}
\arguments{
\item{object}{character vector of graph6 symbols}

\item{...}{other arguments passed to \code{\link[igraph:graph_from_adjacency_matrix]{igraph::graph_from_adjacency_matrix()}}}
}
\value{
A list of igraph objects.
}
\description{
Convert graph6 symbols to igraph objects
}
