% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_from_sparse6.R
\name{igraph_from_sparse6}
\alias{igraph_from_sparse6}
\title{Convert sparse6 symbols to igraph objects}
\usage{
igraph_from_sparse6(object, ...)
}
\arguments{
\item{object}{character vector of sparse6 symbols}

\item{...}{other arguments passed to \code{\link[igraph:graph_from_edgelist]{igraph::graph_from_edgelist()}}}
}
\value{
A list of igraph objects.
}
\description{
Convert sparse6 symbols to igraph objects
}
