% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_from_text.R
\name{network_from_text}
\alias{network_from_text}
\title{Create network objects from graph6, sparse6, or digraph6 symbols}
\usage{
network_from_text(object)
}
\arguments{
\item{object}{character vector with a mixture of graph6, sparse6 or digraph6
symbols}
}
\value{
A list of \href{network::network()}{network} objects.
}
\description{
Create network objects from graph6, sparse6, or digraph6 symbols
}
