% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgraph6-package.R
\docType{package}
\name{rgraph6-package}
\alias{rgraph6-package}
\alias{rgraph6}
\title{rgraph6: Representing Graphs as 'graph6', 'digraph6' or 'sparse6' Strings}
\description{
This package implements methods for representing graphs in formats 'graph6',
'sparse6' and 'digraph6' as strings of printable ASCII characters. The
formats are due to \href{http://cs.anu.edu.au/~bdm}{Brendan McKay} and are
described \href{http://cs.anu.edu.au/people/bdm/data/formats.txt}{here}.
}
\details{
The main functions are \code{\link[=as_graph6]{as_graph6()}}, \code{\link[=as_digraph6]{as_digraph6()}}, \code{\link[=as_sparse6]{as_sparse6()}} for
encoding network data and \code{\link[=igraph_from_text]{igraph_from_text()}} and \code{\link[=network_from_text]{network_from_text()}} for
decoding. There are also other low-level functions to decode directly from
\link[=from_dirgaph6]{digraph6}, \link[=from_graph6]{graph6}, and
\link[=from_sparse6]{sparse6}.
}
\section{Authors and citation}{


\strong{Maintainer}: Michal Bojanowski.

\strong{Authors}: Michal Bojanowski
David Schoch.

The formats themselves are by \href{http://cs.anu.edu.au/~bdm}{Brendan McKay}.

When using this package please cite it. Call \code{citation(package="rgraph6")}
to see how.
}

