% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleg6.R
\docType{data}
\name{sampleg6}
\alias{sampleg6}
\title{A sample vector of g6 codes}
\format{
The format is: \verb{chr [1:9] "CR" "CJ" "CN" ...}
}
\usage{
sampleg6
}
\description{
A vector of graph6 symbols each representing an undirected graph.
}
\examples{
data(sampleg6)
l <- as_adjacency(sampleg6)
# list of adjacency matrices
str(l)
}
\keyword{datasets}
