% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_adjacency.R
\name{as_adjacency}
\alias{as_adjacency}
\alias{as_adjacency.default}
\alias{as_adjacency.graph6}
\title{Convert graph6 symbols to adjacency matrices}
\usage{
as_adjacency(object)

\method{as_adjacency}{default}(object)

\method{as_adjacency}{graph6}(object)
}
\arguments{
\item{object}{character vector of graph6 symbols}
}
\value{
A list of length \code{length(object)} of square symmetric
adjacency matrices corresponding to input graph6 symbols.
}
\description{
This function takes a vector of graph6 symbols and converts them into
symmetric square adjacency matrices.
}
\references{
\url{http://cs.anu.edu.au/people/bdm/data/formats.txt}
}
