% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dgraph6.R
\name{as_dgraph6}
\alias{as_dgraph6}
\alias{as_dgraph6.default}
\alias{as_dgraph6.matrix}
\alias{as_dgraph6.list}
\alias{as_dgraph6.igraph}
\alias{as_dgraph6.network}
\alias{print.dgraph6}
\title{Convert adjacency matrix to a dgraph6 symbol}
\usage{
as_dgraph6(object)

\method{as_dgraph6}{default}(object)

\method{as_dgraph6}{matrix}(object)

\method{as_dgraph6}{list}(object)

\method{as_dgraph6}{igraph}(object)

\method{as_dgraph6}{network}(object)

\method{print}{dgraph6}(x, ...)
}
\arguments{
\item{object}{a square binary adjacency matrix or a list of thereof}

\item{x}{a vector of graph6 symbols (of class "dgraph6")}

\item{...}{other arguments

See \code{\link{rgraph6}} for dgraph6 format description.}
}
\value{
A vector of class \code{dgraph6} extending \code{character} with dgraph6 symbols.
}
\description{
This function converts a graph(s) as adjacency matrices to a dgraph6 symbol(s).
}
