% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_digraph6.R
\name{as_digraph6}
\alias{as_digraph6}
\alias{as_digraph6.default}
\alias{as_digraph6.matrix}
\alias{as_digraph6.list}
\alias{as_digraph6.igraph}
\alias{as_digraph6.network}
\title{Convert adjacency matrix to digraph6 symbols}
\usage{
as_digraph6(object)

\method{as_digraph6}{default}(object)

\method{as_digraph6}{matrix}(object)

\method{as_digraph6}{list}(object)

\method{as_digraph6}{igraph}(object)

\method{as_digraph6}{network}(object)
}
\arguments{
\item{object}{a square binary adjacency matrix or a list of thereof

See \link{rgraph6} for digraph6 format description.}
}
\value{
A character vector of digraph6 symbols.
}
\description{
This function converts a graph(s) as adjacency matrices to a digraph6 symbol(s).
}
