% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_graph6.R
\name{as_graph6}
\alias{as_graph6}
\alias{as_graph6.default}
\alias{as_graph6.matrix}
\alias{print.graph6}
\title{Convert adjacency matrix to a graph6 symbol}
\usage{
as_graph6(object)

\method{as_graph6}{default}(object)

\method{as_graph6}{matrix}(object)

\method{print}{graph6}(x, ...)
}
\arguments{
\item{object}{a square adjacency matrix containing 0s and 1s}

\item{x}{a vector of graph6 symbols (of class "graph6")}

\item{...}{other arguments}
}
\value{
A graph6 symbol.
}
\description{
This function converts a graph adjacency matrix to a graph6 symbol.
}
\details{
If necessary, more details than the description above
}
\note{
This function aim at implementing format and methods originally
created by Brendan McKay, \email{bdm@cs.anu.edu.au} and
\url{http://cs.anu.edu.au/~bdm}.
}
