% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_graph6.R
\name{as_graph6}
\alias{as_graph6}
\alias{as_graph6.default}
\alias{as_graph6.matrix}
\alias{as_graph6.list}
\alias{as_graph6.igraph}
\alias{as_graph6.network}
\title{Convert undirected graphs to graph6 symbols}
\usage{
as_graph6(object)

\method{as_graph6}{default}(object)

\method{as_graph6}{matrix}(object)

\method{as_graph6}{list}(object)

\method{as_graph6}{igraph}(object)

\method{as_graph6}{network}(object)
}
\arguments{
\item{object}{a graph or a list of graphs}
}
\value{
A character vector of graph6 symbols.
}
\description{
This function converts undirected graphs to graph6 symbols. Implemented
methods expect the graph to be an adjacency matrix, an igraph, or a network
object.
}
\details{
See \link{rgraph6} for graph6 format description.
}
