% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgraph6-package.R
\docType{package}
\name{rgraph6-package}
\alias{rgraph6-package}
\alias{_PACKAGE}
\alias{rgraph6}
\title{rgraph6: Representing Graphs as 'graph6', 'digraph6' or 'sparse6' Strings}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Encode network data as strings of printable ASCII characters. Implemented methods include formats 'graph6', 'sparse6', and 'digraph6'.
}
\details{
Formats 'graph6', 'sparse6' and 'digraph6' represent graphs as
strings of printable ASCII characters. The formats are due to \href{http://cs.anu.edu.au/~bdm}{Brendan McKay} and are described
\href{http://cs.anu.edu.au/people/bdm/data/formats.txt}{here}.

The main functions are \code{\link[=as_graph6]{as_graph6()}}, \code{\link[=as_digraph6]{as_digraph6()}}, \code{\link[=as_sparse6]{as_sparse6()}} for
encoding network data and \code{\link[=igraph_from_text]{igraph_from_text()}} and \code{\link[=network_from_text]{network_from_text()}} for
decoding. There are also other low-level functions to decode directly from
\link[=from_digraph6]{digraph6}, \link[=from_graph6]{graph6}, and
\link[=from_sparse6]{sparse6}.
}
\section{Citation}{


The formats themselves are by \href{http://cs.anu.edu.au/~bdm}{Brendan McKay}.

When using this package please cite it by referring to:

Bojanowski M, Schoch D (2021). \emph{rgraph6: Representing Graphs as graph6,
dgraph6 or sparse6 Strings}. R package version: 2.0-0, <URL:
https://mbojan.github.com/rgraph6>.

Call \code{citation(package="rgraph6")} for more details and the BibTeX entry.
}

\author{
\strong{Maintainer}: Michal Bojanowski \email{michal2992@gmail.com} (\href{https://orcid.org/0000-0001-7503-852X}{ORCID}) (Kozminski University)

Authors:
\itemize{
  \item David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID}) (Machester University)
}

}
\keyword{package}
